Package refactorings
