+++ gxawsserverless/src/main/java/com/genexus/cloud/serverless/model/EventMessageProperty.java
@@ -1,4 +1,4 @@
-package com.genexus.cloud.serverless.aws;
+package com.genexus.cloud.serverless.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
