+++ gxawsserverless/src/main/java/com/genexus/cloud/serverless/model/EventMessageResponse.java
@@ -1,4 +1,4 @@
-package com.genexus.cloud.serverless;
+package com.genexus.cloud.serverless.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
